requirements.dev.txt
asgiref
astroid
autopep8
certifi
cffi
chardet
cloudinary
coreapi
coreschema
cryptography
defusedxml
Django
django-cors-headers
django-datetimepicker
django-environ
django-filter
django-money
django-safedelete
djangorestframework
djangorestframework-queryfields
djangorestframework-simplejwt
django-storages
drf-jwt
ecdsa
google-api-python-client
gunicorn
httplib2
idna
importlib-metadata
isort
itypes
Jinja2
jsonpickle
lazy-object-proxy
Markdown
MarkupSafe
mccabe
oauth2client
oauthlib
pep8
Pillow
psycopg2-binary
py-moneyed
pyasn1
pyasn1-modules
pycodestyle
pycparser
PyJWT
pylint
pylint-django
pylint-plugin-utils
pyparsing
python-dotenv
python-jose
python3-openid
pytz
requests
requests-oauthlib
rsa
six
social-auth-app-django
social-auth-core
sqlparse
toml
typed-ast
uritemplate
urllib3
wrapt
zipp
